Move `read_json_file` to `util`
